tokio-ai/dashboard-api/mcp-core/requirements.txt
mcp==1.0.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
google-generativeai==0.3.2
requests==2.31.0
python-json-logger==2.0.7
python-dotenv==1.0.0
